% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esetPlot-class.R
\docType{class}
\name{characterORexpressionOrCall-class}
\alias{characterORexpressionOrCall-class}
\title{S4 Class Union with character/expression/call}
\description{
This is used for the definition of the title/axes labels for the ggplot2 version
}
\keyword{internal}
