% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esetPlot-class.R
\docType{class}
\name{esetPlotInteractive-class}
\alias{esetPlotInteractive-class}
\alias{esetPlotInteractive}
\title{a S4 class to represent interactive plots}
\description{
a S4 class to represent interactive plots
}
\section{Slots}{

\describe{
\item{\code{includeTooltip}}{logical, if TRUE, add hoover functionality showing
sample annotation (variables used in the plot) in the plot}

\item{\code{tooltipVars}}{name of extra phenotypic variable(s) 
to add in rbokehEsetPlot to label the samples}

\item{\code{sizePlot}}{vector containing the size of the interactive plot, 
as [width, height], by default: c(600, 400).}
}}

\author{
Laure Cougnaud
}
