% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{formatManualScale}
\alias{formatManualScale}
\title{extend manual scale values if required}
\usage{
formatManualScale(x, valVar, nameVar)
}
\arguments{
\item{x}{data.frame with \code{nameVar}}

\item{valVar}{fixed value of variable of aesthetic}

\item{nameVar}{name of variable for aesthetic}
}
\value{
vector of manual scales
}
\description{
extend manual scale values if required
}
\author{
Laure Cougnaud
}
\keyword{internal}
