% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{formatOutput}
\alias{formatOutput}
\title{format output of \link{plotEset} function}
\usage{
formatOutput(res, object, type, returnEsetPlot)
}
\arguments{
\item{res}{result of specific \link{plotEset} function}

\item{object}{\code{esetPlot} object or extended class}

\item{type}{string type of plot}

\item{returnEsetPlot}{logical, should the object be returned in the output function?}
}
\value{
result
}
\description{
format output of \link{plotEset} function
}
\author{
Laure Cougnaud
}
\keyword{internal}
