% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEset.R
\docType{methods}
\name{getAxesLimits}
\alias{getAxesLimits}
\alias{getAxesLimits,esetPlot-method}
\title{generic for get axes limits}
\usage{
getAxesLimits(object)

\S4method{getAxesLimits}{esetPlot}(object)
}
\arguments{
\item{object}{\link{plotEset} object}
}
\value{
matrix with limits for axes: columns \code{x} and \code{y}
}
\description{
generic for get axes limits
}
\author{
Laure Cougnaud
}
\keyword{internal}
