% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopElements.R
\name{getCoordGeneSets}
\alias{getCoordGeneSets}
\title{extract coordinates gene sets}
\usage{
getCoordGeneSets(dataPlotGenes, geneSets, esetUsed, geneSetsVar = NULL)
}
\arguments{
\item{dataPlotGenes}{data.frame with two columns 'X' and 'Y' with coordinates for the genes}

\item{geneSets}{geneSets list of gene sets, e.g. gene pathways, output from the 'getGeneSets' function in MLP
the genes IDs must correspond to the sampleNames in the eset}

\item{esetUsed}{expressionSet (or SummarizedExperiment) object with data}

\item{geneSetsVar}{variable of the featureData used to match the genes contained in geneSets,
most probably ENTREZID, if NULL the featureNames of the eSet are used}
}
\value{
data.frame with two columns 'X' and 'Y' with coordinates for the gene sets
}
\description{
extract coordinates gene sets
}
\author{
Laure Cougnaud

Laure Cougnaud
}

