% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEset.R
\docType{methods}
\name{getDataPlotSamplesWithAnnotation}
\alias{getDataPlotSamplesWithAnnotation}
\alias{getDataPlotSamplesWithAnnotation,esetPlot-method}
\alias{getDataPlotSamplesWithAnnotation,ggvisEsetPlot-method}
\alias{getDataPlotSamplesWithAnnotation,plotlyEsetPlot-method}
\title{get sample data for plot}
\usage{
getDataPlotSamplesWithAnnotation(object)

\S4method{getDataPlotSamplesWithAnnotation}{esetPlot}(object)

\S4method{getDataPlotSamplesWithAnnotation}{ggvisEsetPlot}(object)

\S4method{getDataPlotSamplesWithAnnotation}{plotlyEsetPlot}(object)
}
\arguments{
\item{object}{\link{plotEset} object}
}
\value{
data.frame with 'dataPlotSamples' binded
with variables displayed in the plot
}
\description{
get sample data for plot
}
\author{
Laure Cougnaud
}
\keyword{internal}
