% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethodsInputObjectEsetVis.R
\name{getMethodsInputObjectEsetVis}
\alias{getMethodsInputObjectEsetVis}
\title{wrapper to extract useful functions, depending if the object is 
an ExpressionSet or a SummarizedExperiment.}
\usage{
getMethodsInputObjectEsetVis(x)
}
\arguments{
\item{x}{object}
}
\value{
if the object is an ExpressionSet or a SummarizedExperiment, 
returns a list with the functions specific of the class of \code{x},
and equivalent of the ExpressionSet functions: 'sampleNames', 'featureNames',
'fData', 'pData', 'exprs'
\itemize{
	 \item{sampleNames: }{sample names}
  \item{featureNames: }{feature names}
  \item{fData: }{feature annotation}
  \item{pData: }{sample annotation}
  \item{exprs: }{data matrix}
  \item{varLabels: }{sample annotation variables}
  \item{fvarLabels: }{feature annotation variables}
}
}
\description{
This returns an error is \code{x} is not of the correct class.
The package \code{SummarizedExperiment} should be available if \code{x} is of class \code{SummarizedExperiment}.
}
\author{
Laure Cougnaud
}

