% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEset-ggplot.R
\name{ggPlotEset}
\alias{ggPlotEset}
\title{visualize and \link{esetPlot} with the 'ggplot2' package}
\usage{
ggPlotEset(object)
}
\arguments{
\item{object}{object of class \link{esetPlot}}
}
\value{
\code{ggplot} object
}
\description{
visualize and \link{esetPlot} with the 'ggplot2' package
}
\author{
Laure Cougnaud
}
\keyword{internal}
