% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esetPlot-class.R
\docType{class}
\name{ggplotEsetPlot-class}
\alias{ggplotEsetPlot-class}
\alias{ggplotEsetPlot}
\title{a S4 class to represent \code{ggplot} plots}
\arguments{
\item{returnTopElements}{logical, if TRUE (FALSE by default) 
return the outlying elements labelled in the plot (if any)}
}
\description{
a S4 class to represent \code{ggplot} plots
}
\author{
Laure Cougnaud
}
