% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esetPlot-class.R
\docType{class}
\name{ggvisEsetPlot-class}
\alias{ggvisEsetPlot-class}
\alias{ggvisEsetPlot}
\title{a S4 class for \code{ggvis} plot}
\value{
S4 object of class \code{ggvisEsetPlot}
}
\description{
a S4 class for \code{ggvis} plot
}
\section{Slots}{

\describe{
\item{\code{adjustLegend}}{logical, if TRUE (by default) adjust the legends in \code{ggvis} to avoid
overlapping legends when multiple legends}

\item{\code{alphaVar}}{name of numeric variable (in varLabels of the \code{eset}) 
used for the transparency, empty by default.}

\item{\code{alphaRange}}{transparency (alpha) range used in the plot, 
c(0.1, 1) by default.}
}}

\author{
Laure Cougnaud
}
