% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEset-ggvis.R
\name{ggvisPlotEset}
\alias{ggvisPlotEset}
\title{visualize and \link{esetPlot} with the the 'ggvis' package}
\usage{
ggvisPlotEset(object)
}
\arguments{
\item{object}{object of class \link{esetPlot}}
}
\value{
\code{ggvis} plot object
}
\description{
visualize and \link{esetPlot} with the the 'ggvis' package
}
\author{
Laure Cougnaud
}
\keyword{internal}
