% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEset.R
\docType{methods}
\name{plotEset}
\alias{plotEset}
\alias{plotEset,ggplotEsetPlot-method}
\alias{plotEset,ggvisEsetPlot-method}
\alias{plotEset,plotlyEsetPlot-method}
\title{plot an \link{plotEset} object}
\usage{
plotEset(object, returnEsetPlot = FALSE)

\S4method{plotEset}{ggplotEsetPlot}(object, returnEsetPlot = FALSE)

\S4method{plotEset}{ggvisEsetPlot}(object, returnEsetPlot = FALSE)

\S4method{plotEset}{plotlyEsetPlot}(object, returnEsetPlot = FALSE)
}
\arguments{
\item{object}{object of class \link{esetPlot}}

\item{returnEsetPlot}{logical, 
if TRUE return also the \link{esetPlot} object,
such as can be re-use for future call to \link{plotEset}}
}
\value{
the plot object if \code{returnEsetPlot} is FALSE,
otherwise a list with 'plot': the plot object
 and 'esetPlot': the \link{esetPlot} object
}
\description{
plot an \link{plotEset} object
}
\author{
Laure Cougnaud
}
\keyword{internal}
