% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esetPlot-class.R
\docType{class}
\name{plotlyEsetPlot-class}
\alias{plotlyEsetPlot-class}
\alias{plotlyEsetPlot}
\title{a S4 class to represent \code{plotly} plots}
\value{
S4 object of class \code{plotlyEsetPlot}
}
\description{
a S4 class to represent \code{plotly} plots
}
\section{Slots}{

\describe{
\item{\code{returnTopElements}}{logical, if TRUE (FALSE by default)
return the outlying elements labelled in the plot (if any)}

\item{\code{size}}{specified size(s) (cex) for the points, replicated if
needed, used only if \code{sizeVar} is empty, a factor or character
by default: '20' if \code{sizeVar} is not specified}
}}

\author{
Laure Cougnaud
}
\keyword{internal}
