% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEset-plotly.R
\name{plotlyPlotEset}
\alias{plotlyPlotEset}
\title{visualize and \link{esetPlot} with the the 'plotly' package}
\usage{
plotlyPlotEset(object)
}
\arguments{
\item{object}{object of class \link{esetPlot}}
}
\value{
\code{plotly} plot
}
\description{
visualize and \link{esetPlot} with the the 'plotly' package
}
\author{
Laure Cougnaud
}
\keyword{internal}
