% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esetPlot-class.R
\docType{class}
\name{rbokehEsetPlot-class}
\alias{rbokehEsetPlot-class}
\alias{rbokehEsetPlot}
\title{a S4 class to represent \code{rbokeh} plots}
\description{
a S4 class to represent \code{rbokeh} plots
}
\author{
Laure Cougnaud
}
