% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEset-rbokeh.R
\name{rbokehPlotEset}
\alias{rbokehPlotEset}
\title{visualize and \link{esetPlot} with the the 'rbokeh' package}
\usage{
rbokehPlotEset(object)
}
\arguments{
\item{object}{object of class \link{esetPlot}}
}
\value{
\code{rbokeh} plot
}
\description{
visualize and \link{esetPlot} with the the 'rbokeh' package
}
\author{
Laure Cougnaud
}
