% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{setFixElement}
\alias{setFixElement}
\title{check if the aesthetic is fixed (e.g. color, shape, size 'palette')}
\usage{
setFixElement(typeVar, valVar)
}
\arguments{
\item{typeVar}{name of variable for aesthetic}

\item{valVar}{fixed value of variable of aesthetic}
}
\value{
logical, if TRUE the element is fixed
}
\description{
check if the aesthetic is fixed (e.g. color, shape, size 'palette')
}
\author{
Laure Cougnaud
}
\keyword{internal}
