% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{setManualScale}
\alias{setManualScale}
\title{check if manual aesthetic should be set}
\usage{
setManualScale(x, typeVar, valVar)
}
\arguments{
\item{x}{data.frame with \code{typeVar}}

\item{typeVar}{name of variable for aesthetic}

\item{valVar}{fixed value of variable of aesthetic}
}
\value{
logical, if TRUE the manual scale should be set
}
\description{
This is the case only if \code{typeVar} and \code{valVar} are specified,
and if the variable is not numeric or integer (doesn't work with ggplot2)
}
\author{
Laure Cougnaud
}
\keyword{internal}
