% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{simpleCap}
\alias{simpleCap}
\title{capitalize the first letter of a word}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{string}
}
\value{
string with first letter capitalized
}
\description{
capitalize the first letter of a word
}
\keyword{internal}
