% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esetPlot-class.R
\docType{class}
\name{ggvisEsetPlot-class}
\alias{ggvisEsetPlot-class}
\alias{ggvisEsetPlot}
\title{a S4 for \code{ggvis} plot}
\description{
a S4 for \code{ggvis} plot
}
\section{Slots}{

\describe{
\item{\code{adjustLegend}}{logical, if TRUE (by default) adjust the legends in \code{ggvis} to avoid
overlapping legends when multiple legends}
}}

\author{
Laure Cougnaud
}
